Add create/delete medicament + correction affichage role page de profile

+++ app/src/main/java/com/example/gsb/data/model/Categorie.java
@@ -1,15 +1,15 @@
 package com.example.gsb.data.model;
 
 public class Categorie {
-    private Long idCategorie;
+    private int idCategorie;
     private String nameCategorie;
 
-    public Categorie(Long idCategorie, String nameCategorie) {
+    public Categorie(int idCategorie, String nameCategorie) {
         this.idCategorie = idCategorie;
         this.nameCategorie = nameCategorie;
     }
 
-    public Long getIdCategorie() {
+    public int getIdCategorie() {
         return idCategorie;
     }
